Mid-Patch during program design outline
Most of the program design is largely down. The last main question to
ask is how ui components should receive generic data from given rules
and draw these to the screen.

For now, we circumvent this by only focusing on a single rule and
limited ui's, but it should be considered eventually.