Update test framework and fix logic errors
We put in place a little more of the boilerplate required to easily test
particular engine functions.

Preview pieces were added to the testing code, which makes playing quite
a lot easier.

The block data Y values are now mirrored so they match the axis used
internally by the field.

Some engine key logic was altered slightly. This is likely the next big
thing to work on, getting the DAS functionality for fluid movement is a
really important step.