Add simple terminal ui version
This terminal version currently utilizes the linux/keyboard.h header for
access to keyboard state. This is not ideal, and alternatively we could
utilize X11 for key events if we are running an x-server.

This has been partially looked at, the main stoppage point currently is
retrieving the current X11 window from the parent process Id. Or, the
larger problem, how do we find the terminal window id that the program
was run in easily?