Add base gravity to game engine
We now apply gravity per frame. This is pretty minor since gravity
doesn't play a huge part in line race.

What is yet to be added is automatic locking combined with gravity.
There are a few more cases to consider here such as refreshing of lock
count and all those other things, so that is omitted in this commit.